app: move the webapp material into a dedicated folder fix #79
