+++ .bowerrc
@@ -1,3 +1,3 @@
 {
-  "directory": "deeposlandia/static/bower/"
+  "directory": "deeposlandia/webapp/static/bower/"
 }
