+++ deeposlandia/webapp/templates/mapillary_demo.html
